Fix unit test failure because of missing error window initialization
Since the error window is not supposed to exist during a unit test, the BlinkManager was replaced by two events on the main window which may or may not be subscribed to by
the error window.
If it exists, this leads to the same behavior as before, and if it does not exist,
this does not crash.